+++ packages/schematics/src/new/index.ts
@@ -54,7 +54,7 @@ export function m2New(options: m2NewOptions): Rule {
   "private": true,
   "dependencies": {
     "@m2c2kit/addons": "0.3.10",
-    "@m2c2kit/core": "0.3.12"
+    "@m2c2kit/core": "0.3.13"
   },
   "devDependencies": {
     "@m2c2kit/build-helpers": "0.3.9",